backend/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==20.2.0
bcrypt==3.2.0
Brotli==1.0.9
certifi==2024.6.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==2.0.4
constantly==23.10.4
cryptography==41.0.7
daphne==4.1.0
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pip==23.3.1
psycopg2==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==24.0.0
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
service-identity==18.1.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.1.0
wheel==0.41.2
zope.interface==5.4.0
